refactor: update component, file and method naming
